refactor: move nekos.moe code into its own file
